+++ src/parser.cc
@@ -454,7 +454,9 @@ expr_t::parser_t::parse(std::istream& in, const parse_flags_t& flags,
       DEBUG("parser.error", "     token kind = " << int(lookahead.kind));
       DEBUG("parser.error", "   token length = " << lookahead.length);
 
-      add_error_context(line_context(*original_string, pos, end_pos));
+      add_error_context(line_context(*original_string,
+				     static_cast<std::size_t>(pos),
+				     static_cast<std::size_t>(end_pos)));
     }
     throw;
   }
